# Conflicts:
#	BGL/include/CGAL/boost/parameter.h
#	Installation/CHANGES.md
#	Mesh_3/doc/Mesh_3/CGAL/Labeled_mesh_domain_3.h
#	Mesh_3/include/CGAL/Labeled_mesh_domain_3.h

+ fix compilation
+ prepare for using detect_features

+++ Point_set_processing_3/include/CGAL/structure_point_set.h
@@ -19,7 +19,6 @@
 #include <CGAL/disable_warnings.h>
 
 #include <CGAL/property_map.h>
-#include <CGAL/point_set_processing_assertions.h>
 #include <CGAL/assertions.h>
 #include <CGAL/intersections.h>
 
@@ -257,7 +256,7 @@ public:
     {
       m_points.push_back (get(point_map, *it));
       m_normals.push_back (get(normal_map, *it));
-      int plane_index = get (index_map, idx);
+      int plane_index = static_cast<int>(get (index_map, idx));
       if (plane_index != -1)
       {
         m_indices_of_assigned_points[std::size_t(plane_index)].push_back(idx);
